refactor: Расчет эндогенных свойств вынесен в отдельный файл
